+++ src/main/transaction/Transaction.java
@@ -1,13 +1,14 @@
 package transaction;
 
+import java.math.BigDecimal;
 import enums.TransactionType;
 import exception.InsufficientBalanceException;
 
 public abstract class Transaction {
-    protected double amount;
+    protected BigDecimal amount;
     protected TransactionType type;
 
-    public Transaction(double amount, TransactionType type) {
+    public Transaction(BigDecimal amount, TransactionType type) {
         this.amount = amount;
         this.type = type;
     }
