+++ src/main/transaction/TransferTransaction.java
@@ -1,5 +1,6 @@
 package transaction;
 
+import java.math.BigDecimal;
 import enums.TransactionType;
 import model.Account;
 import exception.*;
@@ -8,7 +9,7 @@ public class TransferTransaction extends Transaction {
     private Account sender;
     private Account receiver;
 
-    public TransferTransaction(Account sender, Account receiver, double amount) {
+    public TransferTransaction(Account sender, Account receiver, BigDecimal amount) {
         super(amount, TransactionType.TRANSFER);
         this.sender = sender;
         this.receiver = receiver;
